Add short-circuit min_length() and min_length2() benchmarking.
This adds a test_point.cpp (poorly named, will probably be moved) test that
benchmarks and tests a variety of differnt short-circuting length and length2
comparison methods using a variety of different coordinate systems.

The point of these were to find the most efficient way to implement "find
shortest length/distance" comparisons, because they are all over the code in
lots of hot-spots.