Improve test_point.cpp tests and implementations.
I did a bunch of testing with different ways to implement the Coord methods to
try and optimize things, including alternative lvalue and rvalue
implementations, and various more direct implementations. The conclusions were
a bit inconclusive, with some alternatives seeming to show benefits in some
cases but not others. I suspect they would all end up identical with sufficent
compiler optimization option settings. So in then end I just went with the
simplest solution.

For the main CoordTraits and CoordTraitsBase;

* For change the `En=void` arguments to `typename` instead of `class` as per
  C++ best practices.

* Add a `typename En=void` argument to the CoordTraits template for more
  control over specializing.

* Add `Tmax` and `T2max` attributes for the max range values, limiting `Tmax` to
  `int32_t` max as the upper limit we want to correctly handle.

* Simplify implementation methods and remove commented alternatives.

For the tests;

* Slightly tidy some of the `min_length()` and `min_length2()`
  implementations.

* Change the test data generation to use a fixed seed, and make it generate
  100 different sets of data for the benchmarks. This should make runs more
  consistent and comparable while still avoiding biases in the random data.

* Improve the `checkcoords()` test to check more things to highlight any
  deficiencies in different type options.