Refactor and optimize `Geometry/Bicubic.hpp` to use Eigen better.
This replaces most of the explicit code with smarter use of Eigen matrixes and
vectors. This makes the code both simpler and almost certainly faster.

It also adds a `NumTraits<Matrix<>>` entry so that Matrix and Vector can be
used as the Scalar inside other Matrixs or Vectors. This makes it possible to
do interpolation of vectors as well as scalars.

Note this has not yet been tested, but I wanted to get this in as a starting
point.