Optimize `AABBMesh.cpp` squared_distance() methods.
For some reason there was an excess of vector and index argument copying up
and down the various layers of implementation that I've stripped out. I also
changed some methods to use the standard size_t as the indexed_triange_set
index type instead of int because it makes the method calls pass straight
through the implementation layers.