Rename Fields.hpp's `Parent` to `Base` and improve tests.
The use of `Base` as the typedef for the parent class is more standard in
Eigen. Also add `coef()` and `coefRef()` methods, and add `operator()` for
`v(z,y,x)`.

In the tests simplify the cubic benchmark's way of calculating the point to
interpolate so that the benchmarks better reflect the actual cubic costs. Also
add a commented out cubic benchmark for a vector field, which doesn't yet work.