Add ScalarWrapper.hpp matrix() and array() methods.
These return the unwrapped expression, possibly re-wrapped with
ArrayWrapper or MatrixWrapper if the expression was of the other type.
Note calling matrix() on a Matrix or array() on an Array just returns the
object without wrapping it. There are const and non-const versions of these
methods.

Update the tests to use scalar() and matrix() instead of wrap() and unwrap(),
and remove the comments and macros for wrap() and unwrap().